+++ src/Form/SuccessMessage.js
@@ -2,12 +2,12 @@ import "./styles.css";
 import React from "react";
 import {Alert} from "react-bootstrap";
 
-export default function SuccessMessage() {
+export default function SuccessMessage({stuffType}) {
 
     return (
         <Alert variant="success">
             <Alert.Heading>
-                Zapisano utwór!
+                Zapisano {stuffType}!
             </Alert.Heading>
         </Alert>
     );
